Add public initializers to Language definitions (#1)
The immediate use-case is to reuse the existing Language types in
rewriting documents.

As the initializers are private, it is not possible to create new
values to add an additional value to an existing array, for example,
or to make modifications past mutating existing values from parsing.

I imagine then extended the types to (recursively) build a String
representation of the document.