Allow for rewriting operations to throw errors (#4)
When using GraphQLLanguage to parse and rewrite GraphQL documents, it has come the need to support throwing exceptions when rewrite operations cannot be complete or fail.

To achieve this, both `Rewriter` and `Visitable` protocols have to be updated, together with the methods involved in the process, so exceptions can be thrown up the chain.