Crash when getting `BuildError.debugDescription`
**Problems**

The GraphQL parser's contexts may throw `BuildError.unexpectedContext`
if the context is unable to resolve to its built counterpart. This is
easily replicated by causing an `Int32` overflow. Throwing in these
cases is correct.

By the time the error is consumed, the captured context is no longer
valid. The context is only valid while its associated document exists.
As a result, `context.toStringTree(GraphQLParser.ruleNames)`, which is
called from within the error's computed `debugDescription`, causes the
program to crash.
```
Thread 1: Fatal error: Unexpectedly found nil while unwrapping an
Optional value
```

**Solution**

Instead of capturing the context, `BuildError.unexpectedContext`
captures only what's necessary to generate the error message. This is
calculated before the error is thrown rather than while the error is
consumed, so the context is still valid.